farmlib/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "farmlib"
version = "0.1.0"
description = "OpenClaw Agent Farm SDK — spawn, orchestrate, and analyze multi-agent experiments"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "networkx>=3.2",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]
analysis = ["plotly", "scikit-learn", "sentence-transformers"]

[tool.hatch.build.targets.wheel]
packages = ["farmlib"]

[tool.ruff]
line-length = 100
